Refactor service layer and extract repository logic
