+++ WombatProject/Services/GoogleBooksAPIService.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
+﻿using Microsoft.Extensions.Configuration;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using System;
@@ -14,14 +13,12 @@ namespace WombatLibrarianApi.Services
     public class GoogleBooksAPIService : IBookAPIService
     {
         private readonly IConfiguration _configuration;
-        public WombatBooksContext Context { get; }
-        public List<Book> AuthorBookItems { get; set; } = new List<Book>();
-        public List<Book> SearchResults { get; set; } = new List<Book>();
+        public List<Book> AuthorBookItems { get; } = new List<Book>();
+        public List<Book> SearchResults { get; } = new List<Book>();
 
-        public GoogleBooksAPIService(IConfiguration configuration, WombatBooksContext context)
+        public GoogleBooksAPIService(IConfiguration configuration)
         {
             _configuration = configuration;
-            Context = context;
         }
 
         public async Task GetSearchResults(string searchTerm)
@@ -31,7 +28,7 @@ namespace WombatLibrarianApi.Services
             IList<JToken> tokens = await GetBookItemsAsJToken(url);
             foreach (JToken token in tokens)
             {
-                SearchResults.Add(parseJsonToken(token));
+                SearchResults.Add(ParseJsonToken(token));
             }
         }
 
@@ -42,52 +39,8 @@ namespace WombatLibrarianApi.Services
             IList<JToken> tokens = await GetBookItemsAsJToken(url);
             foreach (JToken token in tokens)
             {
-                AuthorBookItems.Add(parseJsonToken(token));
+                AuthorBookItems.Add(ParseJsonToken(token));
             }
-        }
-
-        public async Task<IEnumerable<object>> GetBooksFromBookshelf()
-        {
-            var bookIds = Context.Bookshelves.Select(book => book.BookId).ToList();
-            return await Context.Books
-                .Where(book => bookIds.Contains(book.Id))
-                .Include(bookShelfItem => bookShelfItem.Authors)
-                .Include(bookShelfItem => bookShelfItem.Categories)
-                .Join(Context.Bookshelves,
-                book => book.Id,
-                bookshelf => bookshelf.BookId,
-                (book, bookshelf) => new 
-                        {
-                            Id = book.Id,
-                            Title = book.Title,
-                            Subtitle = book.Subtitle,
-                            Thumbnail = book.Thumbnail,
-                            Description = book.Description,
-                            PageCount = book.PageCount,
-                            Rating = book.Rating,
-                            RatingCount = book.RatingCount,
-                            Language = book.Language,
-                            MaturityRating = book.MaturityRating,
-                            Published = book.Published,
-                            Publisher = book.Publisher,
-                            BookshelfId = bookshelf.Id
-                        })
-                .ToListAsync();
-        }
-
-        public async Task<Bookshelf> AddBookToBookshelf(Book book)
-        {
-            var bookItem = Context.Books.Where(item => item.Id == book.Id).FirstOrDefault();
-            if (bookItem == null) {
-                Context.Authors.AddRange(book.Authors);
-                Context.Categories.AddRange(book.Categories);
-                Context.Books.Add(book);
-            }
-            Bookshelf bookshelf = new Bookshelf() { BookId = book.Id };
-            Context.Bookshelves.Add(bookshelf);
-            await Context.SaveChangesAsync();
-            return bookshelf;
-
         }
 
         private async Task<IList<JToken>> GetBookItemsAsJToken(string url)
@@ -110,7 +63,7 @@ namespace WombatLibrarianApi.Services
 
         }
 
-        private Book parseJsonToken(JToken jToken)
+        private Book ParseJsonToken(JToken jToken)
         {
 
             JObject volumeInfo = (JObject)jToken["volumeInfo"];
@@ -139,51 +92,6 @@ namespace WombatLibrarianApi.Services
             book.Published = volumeInfo["publishedDate"]?.ToString();
             book.Publisher = volumeInfo["publisher"]?.ToString();
             return book;
-        }
-
-        public async Task<IEnumerable<object>> GetBooksFromWishlist()
-        {
-            var bookIds = Context.Wishlists.Select(book => book.BookId).ToList();
-            return await Context.Books
-                .Where(book => bookIds.Contains(book.Id))
-                .Include(wishlistItem => wishlistItem.Authors)
-                .Include(wishlistItem => wishlistItem.Categories)
-                .Join(Context.Wishlists,
-                book => book.Id,
-                wishlist => wishlist.BookId,
-                (book, wishlist) => new
-                {
-                    Id = book.Id,
-                    Title = book.Title,
-                    Subtitle = book.Subtitle,
-                    Thumbnail = book.Thumbnail,
-                    Description = book.Description,
-                    PageCount = book.PageCount,
-                    Rating = book.Rating,
-                    RatingCount = book.RatingCount,
-                    Language = book.Language,
-                    MaturityRating = book.MaturityRating,
-                    Published = book.Published,
-                    Publisher = book.Publisher,
-                    WishlistId = wishlist.Id
-                })
-                .ToListAsync();
-        }
-
-        public async Task<Wishlist> AddBookToWishlist(Book book)
-        {
-            var bookItem = Context.Books.Where(item => item.Id == book.Id).FirstOrDefault();
-
-            if (bookItem == null)
-            {
-                Context.Authors.AddRange(book.Authors);
-                Context.Categories.AddRange(book.Categories);
-                Context.Books.Add(book);
-            }
-            Wishlist wishlist = new Wishlist() { BookId = book.Id };
-            Context.Wishlists.Add(wishlist);
-            await Context.SaveChangesAsync();
-            return wishlist;
         }
     }
 }
